+++ projects/Mallard/src/screens/article-screen.tsx
@@ -32,7 +32,7 @@ import {
     ArticleNavigationProps,
     ArticleRequiredNavigationProps,
     getArticleNavigationProps,
-} from 'src/navigation/helpers'
+} from 'src/navigation/helpers/base'
 import { routeNames } from 'src/navigation/routes'
 import { color } from 'src/theme/color'
 import { metrics } from 'src/theme/spacing'
